+++ portal/src/main/java/edu/isi/wings/portal/resources/DataResource.java
@@ -173,7 +173,7 @@ public class DataResource extends WingsResource {
     @FormParam("data_type") String data_type,
     @FormParam("props_json") String json
   ) {
-    if (this.dc != null && this.isOwner() && !config.isSandboxed()) {
+    if (this.dc != null && this.isOwner() && !config.portalConfig.isSandboxed()) {
       RunController.invalidateCachedAPIs();
       return this.dc.saveDatatypeJSON(data_type, json);
     }
@@ -190,7 +190,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.addDatatype(parent_type, data_type)
     ) {
       RunController.invalidateCachedAPIs();
@@ -208,7 +208,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.delDatatypes(data_types)
     ) {
       RunController.invalidateCachedAPIs();
@@ -228,7 +228,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.moveDatatypeTo(data_type, from_parent_type, to_parent_type)
     ) {
       RunController.invalidateCachedAPIs();
@@ -248,7 +248,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.moveDataTo(data_id, from_parent_type, to_parent_type)
     ) {
       RunController.invalidateCachedAPIs();
@@ -267,7 +267,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.addDataForDatatype(data_id, data_type)
     ) {
       RunController.invalidateCachedAPIs();
@@ -283,7 +283,7 @@ public class DataResource extends WingsResource {
     @FormParam("data_url") String data_url,
     @FormParam("data_type") String data_type
   ) {
-    if (this.dc != null && this.isOwner() && !config.isSandboxed()) {
+    if (this.dc != null && this.isOwner() && !config.portalConfig.isSandboxed()) {
       RunController.invalidateCachedAPIs();
       return this.dc.addRemoteDataForType(data_url, data_type);
     }
@@ -297,7 +297,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.delData(data_id)
     ) {
       RunController.invalidateCachedAPIs();
@@ -316,7 +316,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.renameData(data_id, newid)
     ) {
       RunController.invalidateCachedAPIs();
@@ -335,7 +335,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.renameDataType(data_type, newid)
     ) {
       RunController.invalidateCachedAPIs();
@@ -354,7 +354,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.setDataLocation(data_id, location)
     ) {
       RunController.invalidateCachedAPIs();
@@ -377,7 +377,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.addBatchData(
           data_type,
           data_ids,
@@ -400,7 +400,7 @@ public class DataResource extends WingsResource {
     @FormParam("newname") String newname,
     @FormParam("metadata_json") String json
   ) {
-    if (this.dc != null && this.isOwner() && !config.isSandboxed()) {
+    if (this.dc != null && this.isOwner() && !config.portalConfig.isSandboxed()) {
       RunController.invalidateCachedAPIs();
       return this.dc.registerData(data_id, newname, json);
     }
@@ -419,7 +419,7 @@ public class DataResource extends WingsResource {
     if (
       this.dc != null &&
       this.isOwner() &&
-      !config.isSandboxed() &&
+      !config.portalConfig.isSandboxed() &&
       this.dc.importFromExternalCatalog(data_id, data_type, json, location)
     ) {
       RunController.invalidateCachedAPIs();
